requirements.txt
fastapi
uvicorn
httpx[http2]
aiohttp
redis
pydantic
python-dotenv
orjson
